requirements.txt
Flask==3.0.2
duckdb==0.10.2
pandas==2.2.2
openai==0.28.1
matplotlib==3.8.4
python-dotenv==1.0.1
reportlab==4.2.0
markdown==3.10.2